[FIX] Handle no lease
When table_allocator_client fails to get a lease, it will output 0. If
we see this value, add the routes to the main routing table, using the
interface index as metric. This routing might have unintentional
consequences, for example with regards to default rules, but it will
ensure that the node at least has a valid routing config and should
reach the internet.

Since there is not supposed to be any default routes in the main table,
you might want to implement a watchdog checking for this.