[SRC] Send RELEASE when adress/link is deleted
I had completely forgot to send a release message when the address or
link is deleted, handling release in server was even ready ... After
this commit, the client will send a release when address or link is
deleted. The sending is (currently?) strictly best-effort, as the client
application will stop after the message is sent.