[SRC] Add NetworkManager support
NetworkManager runs the ifup-scripts installed on systems when an
interface goes up or down. Add a path in our ifup-script for handling
NetworkManager up-events. Using ifup means we don't have to deal with
for example resolvconf.

When used together with NetworkManager, we don't delete any rules from
the main table. This is done so that NetworkManager's policies for which
interface to be used as default can be applied. Keeping a default rule
in the main table also means that the rules with priority 91000 does not
have an effect, but I will keep them around for now when NetworkManager
is used (in case something changes).

If you want to use a specific interface, you need to, like before, bind
to the interface/ip. If you want a rule to apply to some special
traffic, you need to make sure it is placed before the rule pointing to
the main table (pref. 32766 by default).